+++ server/src/main/java/org/opensearch/indices/replication/SegmentReplicationTarget.java
@@ -18,7 +18,6 @@ import org.opensearch.action.StepListener;
 import org.opensearch.common.UUIDs;
 import org.opensearch.common.lucene.Lucene;
 import org.opensearch.common.util.CancellableThreads;
-import org.opensearch.index.engine.exec.coord.CatalogSnapshot;
 import org.opensearch.index.shard.IndexShard;
 import org.opensearch.index.store.Store;
 import org.opensearch.index.store.StoreFileMetadata;
@@ -103,7 +102,7 @@ public class SegmentReplicationTarget extends AbstractSegmentReplicationTarget {
                 checkpointInfoResponse.getInfosBytes(),
                 checkpointInfoResponse.getCheckpoint().getSegmentsGen()
             );
-            indexShard.finalizeReplication(CatalogSnapshot.fromSegmentInfos(infos, store.shardFormatDirectoryResolver()));
+            indexShard.finalizeReplication(Store.fromSegmentInfos(infos, store.shardFormatDirectoryResolver()));
         } catch (CorruptIndexException | IndexFormatTooNewException | IndexFormatTooOldException ex) {
             // this is a fatal exception at this stage.
             // this means we transferred files from the remote that have not be checksummed and they are
